// src/app/Models/RMS-Models/ISummitionProgramDTO.ts
export interface ISummitionProgramDTO {
    summitionProgramID: string;
    autoID: number;
    collegeID: string;
    collegeName: string;
    programID: string;
    programName: string;
    prepareStatements: boolean;
    highestApplicationForUniversities: boolean;
    sectorAgreement: boolean;
    agreementOfHighestUniversitie: boolean;
    agreementPrivateUniversity: boolean;
}